+++ Citadel-Coder/src/utopia/citadel/coder/model/scala/ScalaType.scala
@@ -4,6 +4,13 @@ import scala.language.implicitConversions
 
 object ScalaType
 {
+	// ATTRIBUTES   ---------------------------
+	
+	val string = basic("String")
+	val int = basic("Int")
+	val long = basic("Long")
+	
+	
 	// IMPLICIT  ------------------------------
 	
 	// Implicitly converts from a reference
@@ -11,6 +18,12 @@ object ScalaType
 	
 	
 	// OTHER    -------------------------------
+	
+	/**
+	  * @param contentType Option content type
+	  * @return An option type
+	  */
+	def option(contentType: ScalaType) = generic("Option", contentType)
 	
 	/**
 	  * @param name Name of the (basic) data type
@@ -23,6 +36,25 @@ object ScalaType
 	  * @return A data type based on that reference
 	  */
 	def apply(reference: Reference): ScalaType = apply(Right(reference))
+	
+	/**
+	  * Creates a generic type
+	  * @param reference Base type reference
+	  * @param firstParam First type parameter
+	  * @param moreParams More type parameters
+	  * @return A generic type
+	  */
+	def generic(reference: Reference, firstParam: ScalaType, moreParams: ScalaType*) =
+		apply(Right(reference), firstParam +: moreParams.toVector)
+	/**
+	  * Creates a generic type
+	  * @param name Name of the basic generic class
+	  * @param firstParam First type parameter
+	  * @param moreParams More type parameters
+	  * @return A generic type
+	  */
+	def generic(name: String, firstParam: ScalaType, moreParams: ScalaType*) =
+		apply(Left(name), firstParam +: moreParams.toVector)
 }
 
 /**
@@ -30,14 +62,15 @@ object ScalaType
   * @author Mikko Hilpinen
   * @since 30.8.2021, v0.1
   */
-case class ScalaType(data: Either[String, Reference]) extends ScalaConvertible
+case class ScalaType(data: Either[String, Reference], typeParameters: Vector[ScalaType] = Vector())
+	extends ScalaConvertible
 {
 	// COMPUTED ---------------------------------
 	
 	/**
 	  * @return Reference used by this type. None if not referring to any external type.
 	  */
-	def references = data.toOption
+	def references: Set[Reference] = typeParameters.flatMap { _.references }.toSet ++ data.toOption
 	
 	
 	// IMPLEMENTED  -----------------------------